[core]Modularize addN.
